chore(mcp): rename functions to cdp functions
