+++ products/cdp/mcp/cdp_functions.yaml
@@ -8,49 +8,49 @@ category: Functions
 feature: hog_functions
 url_prefix: /pipeline
 tools:
-    environments-functions-list:
+    environments-cdp-functions-list:
         operation: environments_hog_functions_list
         enabled: false
-    environments-functions-create:
+    environments-cdp-functions-create:
         operation: environments_hog_functions_create
         enabled: false
-    environments-functions-retrieve:
+    environments-cdp-functions-retrieve:
         operation: environments_hog_functions_retrieve
         enabled: false
-    environments-functions-update:
+    environments-cdp-functions-update:
         operation: environments_hog_functions_update
         enabled: false
-    environments-functions-partial-update:
+    environments-cdp-functions-partial-update:
         operation: environments_hog_functions_partial_update
         enabled: false
-    environments-functions-destroy:
+    environments-cdp-functions-destroy:
         operation: environments_hog_functions_destroy
         enabled: false
-    environments-functions-enable-backfills-create:
+    environments-cdp-functions-enable-backfills-create:
         operation: environments_hog_functions_enable_backfills_create
         enabled: false
-    environments-functions-invocations-create:
+    environments-cdp-functions-invocations-create:
         operation: environments_hog_functions_invocations_create
         enabled: false
-    environments-functions-logs-retrieve:
+    environments-cdp-functions-logs-retrieve:
         operation: environments_hog_functions_logs_retrieve
         enabled: false
-    environments-functions-metrics-retrieve:
+    environments-cdp-functions-metrics-retrieve:
         operation: environments_hog_functions_metrics_retrieve
         enabled: false
-    environments-functions-metrics-totals-retrieve:
+    environments-cdp-functions-metrics-totals-retrieve:
         operation: environments_hog_functions_metrics_totals_retrieve
         enabled: false
-    environments-functions-icon-retrieve:
+    environments-cdp-functions-icon-retrieve:
         operation: environments_hog_functions_icon_retrieve
         enabled: false
-    environments-functions-icons-retrieve:
+    environments-cdp-functions-icons-retrieve:
         operation: environments_hog_functions_icons_retrieve
         enabled: false
-    environments-functions-rearrange-partial-update:
+    environments-cdp-functions-rearrange-partial-update:
         operation: environments_hog_functions_rearrange_partial_update
         enabled: false
-    functions-list:
+    cdp-functions-list:
         operation: hog_functions_list
         enabled: true
         scopes:
@@ -67,7 +67,7 @@ tools:
             (destination, site_destination, internal_destination, source_webhook, warehouse_source_webhook, site_app,
             transformation) and enabled status via query parameters.
         mcp_version: 1
-    functions-create:
+    cdp-functions-create:
         operation: hog_functions_create
         enabled: true
         scopes:
@@ -97,7 +97,7 @@ tools:
             template_id:
                 description: >
                     ID of a HogFunctionTemplate to derive defaults from (code, inputs_schema, icon, name, description).
-                    Use the function-templates-list tool to find available templates.
+                    Use the cdp-function-templates-list tool to find available templates.
             hog:
                 description: >
                     Source code for the function. For most types this is Hog code; for site_destination and site_app
@@ -108,7 +108,7 @@ tools:
                 description: >
                     Execution priority for transformation functions (lower runs first). Only applies to
                     type=transformation. If omitted, the function is appended at the end.
-    functions-retrieve:
+    cdp-functions-retrieve:
         operation: hog_functions_retrieve
         enabled: true
         scopes:
@@ -122,10 +122,10 @@ tools:
             Get a specific function by ID. Returns the full configuration including source code, inputs schema,
             input values (secrets are masked), filters, mappings, masking config, and runtime status.
         mcp_version: 1
-    functions-update:
+    cdp-functions-update:
         operation: hog_functions_update
         enabled: false
-    functions-partial-update:
+    cdp-functions-partial-update:
         operation: hog_functions_partial_update
         enabled: true
         scopes:
@@ -138,7 +138,7 @@ tools:
         description: >
             Partially update a function. Can enable/disable the function, change its name, description, source code,
             inputs, filters, mappings, or masking config. The 'type' field cannot be changed after creation. To delete a
-            function, use the functions-delete tool instead.
+            function, use the cdp-functions-delete tool instead.
         exclude_params:
             - bytecode
             - transpiled
@@ -149,7 +149,7 @@ tools:
         param_overrides:
             enabled:
                 description: Set to true to activate or false to deactivate the function.
-    functions-delete:
+    cdp-functions-delete:
         operation: hog_functions_destroy
         enabled: true
         scopes:
@@ -163,10 +163,10 @@ tools:
             Delete a function by ID (soft delete). The function will no longer appear in lists or process events,
             but historical data is preserved.
         soft_delete: true
-    functions-enable-backfills-create:
+    cdp-functions-enable-backfills-create:
         operation: hog_functions_enable_backfills_create
         enabled: false
-    functions-invocations-create:
+    cdp-functions-invocations-create:
         operation: hog_functions_invocations_create
         enabled: true
         scopes:
@@ -180,22 +180,22 @@ tools:
             Test-invoke a function with a mock event payload. Sends the function configuration and test data to the
             plugin server for execution and returns logs and status. Use 'mock_async_functions: true' (default) to
             simulate external calls like fetch() without making real HTTP requests.
-    functions-logs-retrieve:
+    cdp-functions-logs-retrieve:
         operation: hog_functions_logs_retrieve
         enabled: false
-    functions-metrics-retrieve:
+    cdp-functions-metrics-retrieve:
         operation: hog_functions_metrics_retrieve
         enabled: false
-    functions-metrics-totals-retrieve:
+    cdp-functions-metrics-totals-retrieve:
         operation: hog_functions_metrics_totals_retrieve
         enabled: false
-    functions-icon-retrieve:
+    cdp-functions-icon-retrieve:
         operation: hog_functions_icon_retrieve
         enabled: false
-    functions-icons-retrieve:
+    cdp-functions-icons-retrieve:
         operation: hog_functions_icons_retrieve
         enabled: false
-    functions-rearrange-partial-update:
+    cdp-functions-rearrange-partial-update:
         operation: hog_functions_rearrange_partial_update
         enabled: true
         scopes:
